Improve personalization for users of large sites
Allow callers to specify a `section` name that the classifier can
use to develop a topics list, to improve personalization for users
of large, multi-topic sites.

If the topic list is per-hostname, a user of a large general-interest
site may receive inadequate personalization compared to a user of
multiple niche sites with only a few topics per site.

For example, a classifier might map a large news site's hostname to
the topics "real estate", "political news", and "crossword puzzles"
resulting in sub-optimal personalization for the users who primarily
visit the site for its fashion coverage, electronics reviews,
or parenting advice.

Without section information, a general-interest video hosting or
social site would also provide very little helpful personalization
info to its users, while it receives helpful personalization info
as a result of the same people's visits to small sites that cover
only a few topics.

A section can be any subdivision of a site, including a "channel"
"group" or "space."