The noise doesn't provide k-anon
The text claims that the noise provides k-anonymity, which isn't quite right. It does provide k-anonymity at the global level, but at the individual caller level, which is filtered based on which topics the caller has observed, it's more correct to state that each topic will be returned a minimum fraction of the time.