[spec] Make the epoch introduction delay per-site per-epoch
This increases user privacy by making it more challenging to identify a specific user based on the time between two epoch switches. The typical duration between switches will be 7 days for active users. However, this duration can be longer if the browser is not running at the originally scheduled time, becoming effectively a user-specific value.